# contracts/mathmodel/openapi.yaml
openapi: 3.0.3
info:
  title: Level2 Mathmodel API
  description: |
    Draft API for the Level2-mathmodel monitoring + planning MVP.

    **Boundaries (Wave 0 / MVP):**
    - Reads Level2 Collector via `LEVEL2_API_URL` (read-only client).
    - Owns **local variables**, **tag catalog/bindings** (imported from Level2), and planning stubs.
    - Never writes to PLC / Level2.
    - Separate containers/images from Level2; default host port **8090**.
  version: 0.2.0-draft
  contact:
    name: popelev / Level2-mathmodel
servers:
  - url: http://127.0.0.1:8090
    description: Lab mathmodel (host port 8090)
security: []
...
paths:
...
  /api/v1/imports/level2/tags:
    post:
      tags: [imports]
      summary: Import tags from Level2 into catalog
      description: |
        Read-only fetch via Level2Client (`GET /api/v1/devices`, `GET /api/v1/tags`).
        Upserts `TagCatalogEntry` rows keyed by `tag_id` + `device_id`.
        Requires `LEVEL2_API_URL`. Optional `LEVEL2_API_TOKEN` is forwarded as Bearer / X-API-Token.
      responses:
        "200":
          description: Import counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Level2ImportResult"
        "503":
          description: Level2 unreachable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Level2UnavailableError"
  /api/v1/imports/level2/tags/preview:
    post:
      tags: [imports]
      summary: Preview Level2 tag import (dry-run)
      description: Same fetch/normalize as import, but does not persist catalog.
      responses:
        "200":
          description: Dry-run result with preview rows
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Level2ImportPreview"
        "503":
          description: Level2 unreachable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Level2UnavailableError"
...
  /api/v1/bindings:
    get:
      tags: [bindings]
      summary: List logical tag bindings
      responses:
        "200":
          description: Bindings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TagBinding"
    put:
      tags: [bindings]
      summary: Replace or upsert bindings
      description: |
        Body may be a binding array (replace) or
        `{ "mode": "replace"|"upsert", "bindings": [...] }`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: array
                  items:
                    $ref: "#/components/schemas/TagBinding"
                - $ref: "#/components/schemas/TagBindingsPut"
      responses:
        "200":
          description: Current bindings after write
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TagBinding"
        "400":
          description: Invalid body
  /api/v1/bindings/{logical_name}:
    parameters:
      - name: logical_name
        in: path
        required: true
        schema:
          type: string
    delete:
      tags: [bindings]
      summary: Delete one binding
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
components:
...
  schemas:
...
    TagCatalogEntry:
      type: object
      required: [tag_id, device_id]
      properties:
        tag_id:
          type: string
          description: Stable Level2 tag id
        device_id:
          type: string
        path:
          type: string
        datatype:
          type: string
        enabled:
          type: boolean
        writable:
          type: boolean
        interval_ms:
          type: integer
        node_id:
          type: string
        last_seen_at:
          type: string
          format: date-time
        raw:
          type: object
          additionalProperties: true
          description: Optional debug snapshot from Level2 row
    TagBinding:
      type: object
      required: [logical_name, tag_id]
      properties:
        logical_name:
          type: string
          description: Engine logical name for the bound tag
        tag_id:
          type: string
        device_id:
          type: string
        role:
          type: string
          enum: [input, output]
          default: input
        section_id:
          type: string
          description: Optional copper section reference
        cell_id:
          type: string
          description: Optional copper cell reference
        signal:
          type: string
          description: Optional copper signal name
    TagBindingsPut:
      type: object
      required: [bindings]
      properties:
        mode:
          type: string
          enum: [replace, upsert]
          default: replace
        bindings:
          type: array
          items:
            $ref: "#/components/schemas/TagBinding"
    Level2ImportResult:
      type: object
      required: [dry_run, devices_seen, tags_fetched, inserted, updated, total]
      properties:
        dry_run:
          type: boolean
          example: false
        devices_seen:
          type: integer
        tags_fetched:
          type: integer
        inserted:
          type: integer
        updated:
          type: integer
        total:
          type: integer
    Level2ImportPreview:
      allOf:
        - $ref: "#/components/schemas/Level2ImportResult"
        - type: object
          properties:
            preview:
              type: array
              items:
                $ref: "#/components/schemas/TagCatalogEntry"
    Level2UnavailableError:
      type: object
      required: [error, detail]
      properties:
        error:
          type: string
          example: level2_unavailable
        detail:
          type: string
...